Adjusted the modular structure
- Removed the task and event modules since the logic could all be a part of the main agenda module
- Moved the network module to be under agenda
- Fixed a typo with convention
