+++ agenda/network/.gitignore
