+++ agenda/network/proguard-rules.pro
