+++ agenda/network/src/main/AndroidManifest.xml
