+++ authentication/data/build.gradle.kts
@@ -43,5 +43,4 @@ dependencies {
     implementation(projects.authentication.domain)
     implementation(projects.core.domain)
     implementation(projects.core.data)
-    implementation(projects.core.network)
 }
