--- build-logic/gragle.properties
@@ -1,3 +1,0 @@
-org.gradle.parallel=true
-org.gradle.caching=true
-org.gradle.configureondemand=true
