--- events/domain/src/main/java/com/gharseldin/events/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.gharseldin.events.domain
-
-class MyClass {
-}
