+++ settings.gradle.kts
@@ -29,15 +29,10 @@ include(":authentication:domain")
 include(":agenda:presentation")
 include(":agenda:data")
 include(":agenda:domain")
-include(":events:presentation")
-include(":events:data")
-include(":events:domain")
-include(":tasks:presentation")
-include(":tasks:data")
-include(":tasks:domain")
+include(":agenda:network")
 include(":core:presentation:ui")
 include(":core:presentation:designsystem")
 include(":core:data")
 include(":core:database")
-include(":core:network")
 include(":core:domain")
+
