--- tasks/domain/src/main/java/com/gharseldin/tasks/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.gharseldin.tasks.domain
-
-class MyClass {
-}
